NFC: move C++ matchers to a separate file
They are useful beyond strategies. This prepares for the connection of
C++ matchers to transform dialect callbacks.
