+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/CMakeLists.txt
@@ -26,9 +26,11 @@ iree_cc_library(
   HDRS
     "CommonExtensions.h"
     "CommonExtensionsOps.h.inc"
+    "TransformMatchers.h"
   SRCS
     "CommonExtensions.cpp"
     "CommonExtensionsOps.cpp.inc"
+    "TransformMatchers.cpp"
   DEPS
     ::CommonExtensionsOpGen
     IREEDialectsTransforms
@@ -36,11 +38,14 @@ iree_cc_library(
     IREELinalgTransformDialect
     LLVMSupport
     MLIRAffineDialect
+    MLIRAnalysis
     MLIRArithDialect
     MLIRArithUtils
     MLIRBufferizationDialect
     MLIRBufferizationTransforms
     MLIRGPUOps
+    MLIRIR
+    MLIRLinalgDialect
     MLIRLinalgTransformOps
     MLIRLinalgTransforms
     MLIRMemRefDialect
